requirements.txt
fastapi
uvicorn
python-dotenv
cohere
langchain
pypdf
chromadb